+++ asyncgit/src/sync/commits_info.rs
@@ -10,6 +10,12 @@ use unicode_truncate::UnicodeTruncateStr;
 )]
 pub struct CommitId(Oid);
 
+impl Default for CommitId {
+	fn default() -> Self {
+		Self(Oid::zero())
+	}
+}
+
 impl CommitId {
 	/// create new `CommitId`
 	pub const fn new(id: Oid) -> Self {