+++ asyncgit/src/sync/mod.rs
@@ -27,6 +27,7 @@ mod staging;
 mod stash;
 mod state;
 pub mod status;
+mod submodules;
 mod tags;
 mod tree;
 pub mod utils;
@@ -80,6 +81,9 @@ pub use stash::{
 };
 pub use state::{repo_state, RepoState};
 pub use status::is_workdir_clean;
+pub use submodules::{
+	get_submodules, update_submodule, SubmoduleInfo, SubmoduleStatus,
+};
 pub use tags::{
 	delete_tag, get_tags, get_tags_with_metadata, CommitTags, Tag,
 	TagWithMetadata, Tags,